requirements.txt
streamlit
faiss-cpu
numpy
sentence-transformers
openai
pdfplumber
beautifulsoup4
requests
langchain-text-splitters
